requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
pinecone-client>=3.0.0
transformers>=4.36.0
huggingface_hub>=0.20.0
torch>=2.0.0
